% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{Acins}
\alias{Acins}
\title{Calculate Area of Intersection of Circle and Square with Common Center}
\usage{
Acins(r, s)
}
\arguments{
\item{r}{radius of the circle (vector or scalar)}

\item{s}{half-width of the square (scalar)}
}
\description{
Calculate Area of Intersection of Circle and Square with Common Center
}
\examples{
 # calculate area in annulus intersecting square
 diff(Acins(r = c(inner = 11, outer = 12), s = 10.5))
 # calculate area in series of 1 m annuli extending to corner of square
 s <- 10.5 # radius of square (center to side)
 diff(Acins(r = 0:ceiling(sqrt(2) * s), s))
}
