% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{addCarc}
\alias{addCarc}
\title{Add Carcass Distance Data to Ring Data}
\usage{
addCarc(rdat, cod, unitCol = NULL, scVar = NULL)
}
\arguments{
\item{rdat}{list of data frames with summaries of ring data by turbine. Each
data frame has columns \code{r}, \code{exposure}, \code{pinc}}
}
\description{
Add Carcass Distance Data to Ring Data
}
