% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{aic}
\alias{aic}
\alias{aic.ddArray}
\alias{aic.dd}
\title{Akaike Information Criterion}
\usage{
aic(x, ...)

\method{aic}{ddArray}(x, ...)

\method{aic}{dd}(x, ...)
}
\arguments{
\item{x}{list of models to calculate AICs for}

\item{...}{ignored (but required for S3 functions)}

\item{rCol}{name of the "r" variable in model formulas}

\item{scCol}{name of the search class variable in model formulas}

\item{expoCol}{name of the exposure variable in model formulas}
}
\value{
Data frame with AIC, AICc, deltaAICc, and AICw (AIC for the models
 restricted to within the search radius) for all models in \code{x}
}
\description{
functions for calculating AIC and AICc for carcass dispersion
 models.
}
