% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alt_names}
\alias{alt_names}
\title{Names of the Named Distributions}
\format{
An object of class \code{character} of length 10.
}
\usage{
alt_names
}
\description{
Names of the Named Distributions
}
\keyword{datasets}
