% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{canFit}
\alias{canFit}
\title{Fit an array of candidate distance models}
\usage{
canFit(ringData, rCol = "r", scCol = NULL, expoCol = "exposure",
  ncarcCol = "ncarc", modelSet = "all")
}
\arguments{
\item{ringData}{Data frame with a tally of carcasses within 1m rings, along
with outer radii of the rings, (optional) search class column, and the amount
of area searched in each search class and ring.}

\item{rCol}{name of column in \code{ringData} with outer radius for 1m rings}

\item{scCol}{name of column in \code{ringData} with (optional) subdivisions
of rings into search classes (character strings)}

\item{expoCol}{name of "exposure" column of total area in each ring belonging
to each search class.}

\item{modelSet}{names (vector of character strings) of glm distribution templates
to fit. Default is \code{"all"}, which fits \code{gamma}, \code{lognormal},
\code{logLinear}, \code{logQuadratic}, \code{logCubic}, \code{inverse_gamma},
\code{paranormal_gamma}, \code{Rayleigh}, \code{MaxwellBoltzmann}, \code{constant},
\code{tnormal}, \code{exponential}, \code{Pareto}, \code{chisq}, and
\code{inverse_gaussian}. Any subset of these may also be fit with a single
call to \code{canFit}.}
}
\value{
list of fitted glm models as a \code{dmod} objects, which are standard
 \code{glm} objects with the name of the fitted distribution template attached
 as \code{$distr} and the name of the (optional) search class column name
 attached as \code{$scCol}. In addition, the data in the \code{glm} are
 coded as "ncarc" for number of carcasses, "r" for distance from turbine,
 "exposure" for amount of area in the given search class at the given distance.
}
\description{
Fit glm's for canonical distance models.

Models are fit for glm's corresponding to gamma, lognormal, log-linear,
log-quadratic, log-cubic, truncated normal, exponential, Rayleigh, Pareto,
inverse gamma, paranormal gamma, Maxwell Boltzmann, chi-squared, and inverse
Gaussian distributions. These models apply strictly to the carcasses within
the search radius. Extrapolation outside the search radius requires the
additional step of fitting truncated distributions (\code{tfit.dmod()})
}
