% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carcass_polygon}
\alias{carcass_polygon}
\title{Example Carcass Distances to Accompany the Polygon Layout Data Set}
\format{
A data frame illustrating an import format for carcass distances.
 There are columns with turbine IDs (\code{turbine}) and carcass distances
 from turbine (\code{r}). Distances (\code{r}) are in meters from the nearest
 turbine. The data set is used in the "polygon" example in the User Guide.
}
\usage{
carcass_polygon
}
\description{
Example Carcass Distances to Accompany the Polygon Layout Data Set
}
\keyword{datasets}
