% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carcass_simple}
\alias{carcass_simple}
\title{Carcass Data to Accompany the Simple Geometry Data Format}
\format{
An object of class \code{data.frame} with 55 rows and 2 columns.
}
\usage{
carcass_simple
}
\description{
Carcass Data to Accompany the Simple Geometry Data Format
}
\keyword{datasets}
