% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carcass_simple0}
\alias{carcass_simple0}
\title{Full, Simulated carcass_simple Data Set (Including Locations and Missed Carcasses)}
\format{
An object of class \code{data.frame} with 100 rows and 6 columns.
}
\usage{
carcass_simple0
}
\description{
Full, Simulated carcass_simple Data Set (Including Locations and Missed Carcasses)
}
\keyword{datasets}
