% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{cof2parms.glm}
\alias{cof2parms.glm}
\title{Convert a glm model into a dd object (under construction...)}
\usage{
\method{cof2parms}{glm}(x, rvar = "r", svar = NULL,
  expovar = "exposure", ...)
}
\arguments{
\item{rvar}{name of distance variable}

\item{svar}{name of search class variable}

\item{expovar}{name of exposure variable}

\item{dmod}{a \code{glm} object}
}
\value{
\code{dd} object
}
\description{
Convert a glm model into a dd object (under construction...)
}
