% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{cof2parms}
\alias{cof2parms}
\alias{cof2parms.matrix}
\alias{cof2parms.numeric}
\alias{cof2parms.dd}
\title{Convert glm distance parameters into named distribution parameters}
\usage{
cof2parms(x, ...)

\method{cof2parms}{matrix}(x, distr, ...)

\method{cof2parms}{numeric}(x, distr, ...)

\method{cof2parms}{dd}(x, ...)
}
\arguments{
\item{x}{object (vector or matrix of parameters, dmod, or glm)
with named glm parameters ("r", "I(r^2)", "I(r^3)", "log(r)", or "I(1/r)")}

\item{...}{ignored [but required for CRAN]}

\item{distr}{name of the distribution}
}
\value{
matrix of parameters
}
\description{
Convert glm distance parameters into named distribution parameters
}
