% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cof_name}
\alias{cof_name}
\title{Names of the GLM Coefficients for Each Distribution}
\format{
A list with the names of the coefficients (vector of character strings)
 as they appear in the \code{$coefficients} value returned from \code{glm} for
 each model.
}
\usage{
cof_name
}
\description{
Names of the GLM Coefficients for Each Distribution
}
\keyword{datasets}
