% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{constraints}
\alias{constraints}
\title{Constraints on GLM Coefficients for Extensibility to a Distribution}
\format{
A list of matrices giving the upper and lower bounds that each
 model's coefficients must meet for the model to be extensible to a
 distribution. The \code{parscale} column may be used in \code{optim} for
 fitting a truncated weighted likelihood model.
}
\usage{
constraints
}
\description{
Constraints on GLM Coefficients for Extensibility to a Distribution
}
\keyword{datasets}
