% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{constraints_par}
\alias{constraints_par}
\title{Constraints on Parameters to Assure Extensibility}
\format{
An object of class \code{list} of length 17.
}
\usage{
constraints_par
}
\description{
Constraints on Parameters to Assure Extensibility
}
\keyword{datasets}
