% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{dd2ddSim}
\alias{dd2ddSim}
\title{Extract Parameters from a Distance Model (\code{dd}) and Format as \code{ddSim} Object}
\usage{
dd2ddSim(dd)
}
\arguments{
\item{dd}{\code{dd} object}
}
\value{
\code{ddSim} object with 1 row
}
\description{
This is a utility function called internally by \code{dwp} functions
to extract parameters from a fitted \code{\link[=ddFit]{dd}} model and 
formats them for analysis and calculation.
}
