% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{ddArray}
\alias{ddArray}
\title{Function for fitting glm's and distance distributions to carcass dispersions}
\usage{
ddArray(ringData, scCol = NULL, rCol = "r", expoCol = "exposure",
  ncarcCol = "ncarc", modelSet = "all")
}
\arguments{
\item{ringData}{data frame with a simplified representation of the seach areas
and carcass locations. Columns include the outer radii of concentric rings
around turbine centers (\code{ncarcCol}), the number of carcasses in each
ring (\code{ncarcCol}), an optional column for search class (\code{scCol}),
and the amount of area or "exposure" for each search class in each ring
(\code{expoCol}.}

\item{scCol}{name of the (optional) search class column. Search classes typically
represent ground with different search characteristics resulting in different
detection probabilities. Search classes will affect the number of carcasses
observed (via differences in detection probabilities) but are assumed not to
affect the distribution of distances at which carcasses land after getting
struck by turbines.}

\item{rCol}{name of the column representing the outer radii of concentric rings
for which carcass counts, search exposure, and search class are tallied.}

\item{expoCol}{name of the column representing the amount of area in each ring
and (optional) search class.}

\item{ncarcCol}{name of the column with the number of carcasses observed in
each ring and (optional) search class.}

\item{modelSet}{a vector of names (character strings) with the names of the
distributions to be fit. Default is \code{"all"} which fits \code{gamma},
\code{lognormal}, \code{logLinear}, \code{logQuadratic}, \code{logCubic},
\code{inverse_gamma}, \code{paranormal_gamma}, \code{Rayleigh},
\code{MaxwellBoltzmann}, \code{constant}, \code{tnormal}, \code{exponential},
\code{Pareto}, \code{chisq}, and \code{inverse_gaussian}. Any subset of
these may also be fit with a single call to \code{ddArray}.}
}
\value{
a \code{ddArray} object, which is a list of \code{dd} objects, one for
 each fitted model. Each \code{dd} object consists of the following components:
 \describe{
   \item{\code{$distr}}{name of the fitted distribution}
   \item{code{$parms}}{MLE for parameter vector for the fitted distribution
     (in cases where the model can be extended beyond the search radius) or
     NA (in cases where the model cannot be extended). \code{$parms} is a
     with names corresponding to distribution parameter names (e.g., for a
     truncated normal distribution, the vector would be something like
      \code{c(mean = 28.3, sd = 12.3)}.}
   \item{\code{$beta}}{MLE for parameter vector for the fitted truncated glm
     that may be extended beyond the search radius.}
   \item{\code{$varbeta}}{estimated variance of estimated \code{$beta} vector.
     This variance matrix is used in simulating parameters and generating
     confidence intervals.}
   \item{\code{$ncarc}}{total number of carcasses observed}
   \item{\code{$n}}{number of rings}
   \item{\code{$k}}{number of parameters estimated}
   \item{\code{$llik}}{the log-likelihood for the fitted MLE parameters}
   \item{\code{$dmod}}{\code{dmod} object for the base, non-extendable model,
     applicable only to the carcasses within the search radius. The \code{dmod}
     object itself is a list, which is a \code{glm} object with the name of
     the model as \code{$distr} (e.g., \code{"gamma"} or \code{"lognormal"}).
     The \code{dmod} list has all the usual stuff from a fitted glm, including
     the raw data, covariate information, etc.}
 }
}
\description{
The function first fits the glm models for carcass dispersion within the
search radius. Then it uses truncated glm to fit carcass distributions that
may be extrapolated (with appropriate caution) beyond the search radius. The
fitted models can then be used to estimate dwp or to research carcass dispersion
patterns in general.
}
