% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{ddCI}
\alias{ddCI}
\title{Calculate CI for CDF, PDF, or quantile}
\usage{
ddCI(mod, x, type = "CDF", CL = 0.9, nsim = 1000, extent = "full",
  zrad = 150, na.tol = 0.1)
}
\arguments{
\item{mod}{a \code{dd} object}

\item{x}{distance from turbine (scalar or vector) or probability (for quantile)}

\item{type}{"CDF", "PDF", or "quantile"}

\item{CL}{confidence level for the confidence interval(s)}

\item{nsim}{number of simulation draws to base the estimate of CI on}

\item{na.tol}{maximum fraction of invalid parameter sets to discard when
constructing CIs; abort if \code{mean(mod[, "extensible"]) > na.tol}}
}
\value{
array (\code{ddCI} class) with columns for distance and the CI bounds
}
\description{
Calculate CI for CDF, PDF, or quantile
}
