% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{ddd}
\alias{ddd}
\alias{pdd}
\alias{rcd}
\alias{qdd}
\title{Function to mimic dnorm, dexp, etc. for \code{dd} models}
\usage{
ddd(x, model, extent = "full", zrad = 150)

pdd(x, model, extent = "full", zrad = 150)

rcd(x, model, extent = "full", zrad = 150)

qdd(p, model, extent = "full", zrad = 150, subdiv = 1000)
}
\arguments{
\item{x}{numeric x >= 0}

\item{model}{either a \code{dd} object or the name of a distance model
(e.g., "gamma").}

\item{extent}{"full" or "win" to account for all carcasses or only those
within the search radius.}

\item{parms}{scalar, vector, or array of parameters. Or if \code{x} is a
\code{dd} object, then \code{parms} may be NULL, in which case, \code{ddd}
 uses the MLE parameters found in the \code{dd} object.}

\item{srad}{the radius to use for truncated distributions
(\code{srad = NULL} to use full, non-truncated distribution)}

\item{x}{numeric x >= 0}

\item{model}{either a \code{dd} object or the name of a distance model
(e.g., "gamma").}

\item{parms}{scalar, vector, or array of parameters. Or if \code{x} is a
\code{dd} object, then \code{parms} may be NULL, in which case, \code{ddd}
uses the MLE parameters found in the \code{dd} object.}

\item{x}{numeric x >= 0}

\item{model}{either a \code{dd} object or a set of simulated parameters
(i.e., a \code{ddSim} object)}

\item{srad}{the radius to use for truncated distributions
(\code{srad = NULL} to use full, non-truncated distribution)}
}
\value{
vector of PDF values. If \code{srad = NULL} and a full distribution
 is used, the length of the output is equal to the maximum of the length of
 \code{x} and the number of parameter sets are included in \code{model} (i.e.
 1 if \code{model} is a \code{dd} object, or \code{nrow(model)} if
 \code{model} is a \code{ddSim} object. If a truncated distribution is to be
 used, then \code{length(x) = 1} or the number of parameter sets must be 1
 (which occurs if \code{model} is a \code{dd} object or \code{nrow(model) = 1}
 if \code{model} is a \code{ddSim} object.

answer

vector or array of \code{rcd} values for every parameter set and
 distance. If there is only one parameter set (e.g., if \code{model} is a
 \code{dd} object), then the function returns a vector representing the
 \code{rcd} for each distance in \code{x}. If \code{model} is a \code{ddSim}
 object, then the function returns an array, with each column representing
 the vector \code{rcd(x)} values for all distances for a single parameter set,
 and the rows representing the \code{rcd} values for all parameter sets at a
 given distance. NOTE: \code{PDF(0)/0} is dengenerate in R, and \code{rcd}
 approximates x < 0.001 with x = 0.001 to streamline the calculations.
}
\description{
Function to mimic dnorm, dexp, etc. for \code{dd} models

Function to mimic pnorm, pexp, etc. but

Function to calculate relative carcass densities
}
\details{
\code{rcd} gives relative carcass densities at a point \code{x}
 meters from a turbine. In general, rcd is proportional to PDF(x)/x,
 normalized so that the surface of rotation of rcd(x) has total area of 1.
 There are more stringent contstraints on the allowable parameters in the
 fitted (or simulated) glm's because the integral of PDF(x)/x must converge.
}
