% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{ddSim}
\alias{ddSim}
\alias{ddSim.dd}
\alias{ddSim.dmod}
\title{Simulation of dispersion parameters}
\usage{
ddSim(x, ...)

\method{ddSim}{dd}(x, nsim = 1000, extent = "full", ...)

\method{ddSim}{dmod}(x, nsim = 1000, extent = "full", ...)
}
\arguments{
\item{x}{object to simulate from}

\item{...}{ignored (but required format for CRAN)}

\item{nsim}{number of simulation draws}
}
\value{
array with simulated beta parameters from the glm model, and their
 conversion to distribution parameters
}
\description{
Simulation of dispersion parameters
}
