% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{degOrder}
\alias{degOrder}
\title{An Ordering of the Models by Degree of the Polynomial}
\format{
An object of class \code{character} of length 17.
}
\usage{
degOrder
}
\description{
An Ordering of the Models by Degree of the Polynomial
}
\keyword{datasets}
