% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distr_names}
\alias{distr_names}
\title{Full List of Names of Distributions}
\format{
An object of class \code{character} of length 17.
}
\usage{
distr_names
}
\description{
Full List of Names of Distributions
}
\keyword{datasets}
