% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dmb}
\alias{dmb}
\alias{pmb}
\alias{dxep1}
\alias{pxep1}
\alias{pxep02}
\alias{dxep02}
\alias{dxep12}
\alias{pxep12}
\alias{dxep123}
\alias{pxep123}
\alias{dxepi0}
\alias{pxepi0}
\alias{dxep0123}
\alias{pxep0123}
\alias{dxep012}
\alias{pxep012}
\alias{dxep2}
\alias{pxep2}
\alias{dxep0}
\alias{pxep0}
\title{Calculate Exposure for Corners of Square Plots
Probability Distributions for Carcasses vs. Distance}
\usage{
dmb(x, a)

pmb(q, a)

dxep1(x, b1)

pxep1(q, b1)

pxep02(q, b0, b2)

dxep02(x, b0, b2)

dxep12(x, b1, b2)

pxep12(x, b1, b2)

dxep123(x, b1, b2, b3, const = NULL)

pxep123(x, b1, b2, b3, const = NULL)

dxepi0(x, shape, scale)

pxepi0(x, shape, scale)

dxep0123(x, b0, b1, b2, b3, const = NULL)

pxep0123(x, b0, b1, b2, b3, const = NULL)

dxep012(x, b0, b1, b2, const = NULL)

pxep012(x, b0, b1, b2, const = NULL)

dxep2(x, s2)

pxep2(x, s2)

dxep0(x, a)

pxep0(x, a)
}
\arguments{
\item{x, }{q vector of quantiles}

\item{a, }{b0, b1, b2, b3, shape, scale, s2 parameters used in the respective
distributions. See Detailin the Maxwell-Boltzmann distribution}

\item{b1}{\code{b1} parameter in the log-linear distribution}

\item{b0, }{b1, b2, b3, const parameters in the paranormal-gamma distribution}

\item{shape, }{scale \code{a > 0 (shape), b > 0 (scale)} parameters in the
inverse gamma distribution}

\item{s2}{\code{s2} parameter in the xep2 distribution}

\item{x, }{q vector of quantiles}

\item{x, }{q vector of quantiles}

\item{b1, }{b2 \code{b1, b2} parameters in the log-quadratic distribution}

\item{x, }{q vector of quantiles}

\item{b1, }{b2 \code{b1, b2} parameters in the log-quadratic distribution}

\item{x, }{q vector of quantiles}

\item{x, }{q vector of quantiles}

\item{x, }{q vector of quantiles}

\item{b0, }{b1, b2, const parameters in the xep012 distribution}

\item{x, }{q vector of quantiles}

\item{x, }{q vector of quantiles}

\item{a}{\code{a} parameter in the xep0 distribution with \code{scale = x_m}
fixed at 1.}
}
\value{
\code{dmb} gives the density, \code{pmb} gives the distribution function.
 The length of the result is the maximum of the lengths of the numerical arguments
 for the other functions. The numerical arguments are recycled to the length of
 the result.

\code{dlogL} gives the density, \code{plogL} gives the distribution function.
 The length of the result is the maximum of the lengths of the numerical arguments
 for the other functions. The numerical arguments are recycled to the length of
 the result.

\code{dlogQ} gives the density, \code{plogQ} gives the distribution function.
 The length of the result is the maximum of the lengths of the numerical arguments
 for the other functions. The numerical arguments are recycled to the length of
 the result.

\code{dlogC} gives the density, \code{plogC} gives the distribution
 function. The length of the result is the maximum of the lengths of the
 numerical arguments for the other functions. The numerical arguments are
 recycled to the length of the result.

\code{digam} gives the density, \code{pigam} gives the distribution function.
 The length of the result is the maximum of the lengths of the numerical arguments
 for the otherp functions. The numerical arguments are recycled to the length of
 the result.

\code{dxep0123} gives the density, \code{ppng} gives the distribution function.
 The length of the result is the maximum of the lengths of the numerical arguments
 for the other functions. The numerical arguments are recycled to the length of
 the result.

\code{dxep012} gives the density, \code{pxep012} gives the distribution function.
 The length of the result is the maximum of the lengths of the numerical arguments
 for the other functions. The numerical arguments are recycled to the length of
 the result.

\code{dRay} gives the density, \code{pRay} gives the distribution
 function. The length of the result is the maximum of the lengths of the
 numerical arguments for the other functions. The numerical arguments are
 recycled to the length of the result.

\code{dPare1} gives the density, \code{pPare1} gives the distribution function.
 The length of the result is the maximum of the lengths of the numerical arguments
 for the other functions. The numerical arguments are recycled to the length of
 the result.
}
\description{
Calculate Exposure for Corners of Square Plots
Probability Distributions for Carcasses vs. Distance
}
\details{
The pdf of the log-linear distribution is defined as
 f(x; b1) = b1^2 * x * exp(b1 * x)

The pdf of the log-quadratic distribution is defined as
 f(x; b1, b2) = const * x * exp(b1 * x + b2 * x^2). The normalization constant
 (\code{const}) is a complicated expression but can be expressed using only
 elementary functions and \code{erf}.

The pdf of the log-cubic distribution is defined as
 f(x; b1, b2, b3) = const * x * exp(b1 * x + b2 * x^2 + b3 * x^3). The
 normalization constant (\code{const}) is a complicated expression that
 cannot be expressed using only elementary functions and readily approximated
 functions in R. Calculation of constant is cumbersome, so it may be calculated
 externally and provided. NOTE: The log-cubic distribution is calculated via
 numerical integration. It is vectorized for x but not for b1, b2, b3.

The pdf of the inverse gamma distribution is defined as
 f(x; a, b) = b^a/gamma(a) * x^(-a - 1) * exp(-b/x). Calculation is via
 \code{invgamma::dinvgamma} and \code{invgamma::pinvgamma} with modification
 to return 0 instead of NaN for x <= 0. NOTE: The parameters

The pdf of the paranormal-gamma distribution is defined as
 f(x; b0, b1, b2, b3) = const * x * exp(b0 * log(x) + b1 * x + b2 * x^2 + b3 * x^3).
 Calculation is cumbersome and done via numerical integration and is not
 vectorized yet.

The pdf of the xep012 distribution is defined as
 f(x; b0, b1, b2) = const * x * exp(b0 * log(x) + b1 * x + b2 * x^2).
 Calculation is cumbersome and done via numerical integration and is not
 vectorized (yet).

The pdf of the xep2 distribution is defined as
 f(x; s2) = x/s2 * exp(-x^2/(2 * s2))

The pdf of the Pareto distribution with scale = 1 is defined as
 f(x; a) = a/x^(a + 1)
}
\section{The distributions}{

 Except where otherwise noted, the parameterizations match those given in
 Wikipedia (and the user guide and vignettes)...section under construction.
 \describe{
   \item{Maxwell-Boltzmann}{The pdf of the Maxwell-Boltzmann is defined as
     f(x; a) = sqrt(2/pi)/a^3 * x^2 * exp(-x^2/(2*a^2)), and the offset used in
     the glm for carcass distances is \code{log(exposure) + log(r)}.}
   \item{log-linear}{The pdf of the log-linear is f(x; b1) = b1^2 * x * exp(b1 * x),
     where b1 is the}
 }
}

