% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{erf}
\alias{erf}
\title{Error function}
\usage{
erf(x)
}
\arguments{
\item{x}{numeric scalar or array}
}
\value{
erf(x) with same dimensions as x
}
\description{
The error function is closely related to the standard normal CDF and arises frequently in probability calculations
}
\details{
\eqn{erf(x) = 2/\sqrt\pi * \int_0^x {exp(-x\^2} dx}
}
