% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{estdwp}
\alias{estdwp}
\title{Estimate DWP}
\usage{
estdwp(x, ...)
}
\arguments{
\item{x}{\code{psiHat} object, which is a matrix \code{$psi} giving the
probability of a given carcass landing in the searched area at each turbine.
Column names = turbine IDs.}

\item{...}{ignored}

\item{ncarc}{vector of total carcass count at each turbine represented in x.}

\item{nboot}{number of parametric bootstrap iterations for estimating CIs}

\item{GenEst}{format the results for importing into GenEst}
}
\value{
list
}
\description{
Estimate DWP
}
