% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{estpsi}
\alias{estpsi}
\alias{estpsi.data.frame}
\alias{estpsi.tlist}
\alias{estdwp.psiHat}
\title{Estimate Probability Carcass lands in Searched Area}
\usage{
estpsi(x, ...)

\method{estpsi}{data.frame}(x, model, extent = "full", nsim = 1000,
  zrad = 150, ...)

\method{estpsi}{tlist}(x, model, extent = "full", nsim = 1000,
  zrad = 150, ...)

\method{estdwp}{psiHat}(x, nboot = NULL, ...)
}
\arguments{
\item{x}{data. Either \code{rdat} data frame for a single turbine, site,
or other entity, or a list of such entities.}

\item{...}{ignored}

\item{model}{A fitted \code{dd} model}

\item{extent}{calculate dwp within searched area only (\code{"win"}) or
for fulss complement of carcasses (\code{"full"}).}

\item{nboot}{number of parametric bootstrap iterations for estimating CIs}
}
\value{
Data frame with AIC, AICc, deltaAICc for all models in \code{x}
}
\description{
Estimated probability that carcass lands in searched area. This
 is an intermediate step in estimating dwp but is also interesting in its own
 right.
}
