% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{estpsi.ringscc}
\alias{estpsi.ringscc}
\title{Estimate Probability that Carcass Lands in Searched Area
rpA}
\usage{
\method{estpsi}{ringscc}(x, model, modnames = NULL, extent = "full",
  nsim = 1000, zrad = 150, ...)
}
\arguments{
\item{model}{one selected model for each carcass class or a \code{ddArraycc}
accompanied by a vector of model names for each carcass class}

\item{modnames}{vector of names of model to use for each size class.}
}
\description{
Estimate Probability that Carcass Lands in Searched Area
rpA
}
