% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{exclude}
\alias{exclude}
\title{Remove Particular Names from a Longer List}
\usage{
exclude(what, from)
}
\arguments{
\item{what}{vector of distribution names to exclude}

\item{from}{vector of distribution names to be excluded from}
}
\description{
The intended use is to remove specific distributions from \code{ddArray} for 
subsequent analyses. For example, 
\code{dmod2 <- dmod[exclude("lognormal", names(dmod))]} would subset a list
of models, \code{dmod}, to exclude \code{"lognomal"}.
}
