% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{exportGenEst}
\alias{exportGenEst}
\title{Export Estimated Density-Weighted Proportion to File in Proper GenEst Format}
\usage{
exportGenEst(dwp, file)
}
\arguments{
\item{dwp}{a \code{dwphat} object}

\item{file}{name of file to export the \code{dwp} estimates to}
}
\description{
Export Estimated Density-Weighted Proportion to File in Proper GenEst Format
}
