% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_M.R
\name{fmmax}
\alias{fmmax}
\alias{fmmax.ab}
\title{Find suitable mmax for clipping improper priors for M}
\usage{
fmmax(x, g)

fmmax.ab(x, pBa, pBb)
}
\arguments{
\item{x}{carcass count}

\item{g}{overall carcass detection probability}

\item{pBa, pBb}{parameters for beta distribution characterizing estimated \eqn{g}}
}
\value{
integer \eqn{m} such that \eqn{Pr(M >= m) < 0.0001}
}
\description{
Improper priors need to be clipped in order to be usable.
 \code{fmmax} and \code{fmmax.ab} find values of \eqn{m} that are large enough
 that the probability of exceeding is less than 0.0001 (depends on \eqn{g} and
 \eqn{X}).
}
