% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{formatGenEst}
\alias{formatGenEst}
\title{Format DWP Estimate for Use in GenEst}
\usage{
formatGenEst(dwphat)
}
\arguments{
\item{dwphat}{a \code{dwphat} object}
}
\value{
an \code{nsim*nturbine} by \code{nclass + 1} data frame, with columns
 for the turbine ID and for estimated \code{dwp} for each carcass class (e.g.,
 \code{large}, \code{medium}, \code{small}, \code{bat}).
}
\description{
GenEst requires \code{dwp} data to be formatted as a data frame with columns
 for turbine ID and for estimated \code{dwp} for each carcass class. To
 incorporate uncertainty in the estimates, \code{nsim} simulated copies of
 the basic format are appended to the columns in the data set.
}
