% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{getncarc}
\alias{getncarc}
\alias{getncarc.ringscc}
\alias{getncarc.rings}
\alias{getncarc.xyLayout}
\alias{getncarc.ddArray}
\alias{getncarc.ddArraycc}
\alias{getncarc.dd}
\title{Simple Function to Extract Carcass Counts}
\usage{
getncarc(x, ...)

\method{getncarc}{ringscc}(x, ...)

\method{getncarc}{rings}(x, ...)

\method{getncarc}{xyLayout}(x, ...)

\method{getncarc}{ddArray}(x, ...)

\method{getncarc}{ddArraycc}(x, ...)

\method{getncarc}{dd}(x, ...)
}
\arguments{
\item{x}{a data structure with \code{ncarc} buried in it somewhere}

\item{...}{ignored}
}
\description{
Carcass counts are easy to extract from any of the data structures, but it
 may be difficult to remember where to retrieve the data from for any particular
 structure. \code{getncarc} simplifies the task by having the same usage for
 all data types.
}
