% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{layout_eagle}
\alias{layout_eagle}
\title{Example Bare Vector Format for Eagle Data}
\format{
An object of class \code{data.frame} with 60 rows and 3 columns.
}
\usage{
layout_eagle
}
\description{
Example Bare Vector Format for Eagle Data
}
\keyword{datasets}
