% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{layout_polygon}
\alias{layout_polygon}
\title{Example Polygon Data for Site Layout}
\format{
A data frame illustrating the required raw data format for using
 standard R polygons to characterize a site layout. There must be three
 columns: one giving turbine IDs (\code{turbine}) and columns for the \code{x}
 and \code{y} coordinates that delineate the plot layouts. Turbine IDs must be
 syntactitally valid R names, that is, combinations of letters, numbers,
 underscores ( _ ) and periods ( . ) and no spaces, hyphens, or other special
 characters. Names must not begin with a number, so \code{1, 2, 3, ...},
 \code{3B1}, and \code{.72S} are NOT valid names. Coordinates shoud be in
 meters relative to a turbine at (0, 0).
}
\usage{
layout_polygon
}
\description{
Example Polygon Data for Site Layout
}
\keyword{datasets}
