% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{layout_xy}
\alias{layout_xy}
\title{Example Data for Site Layout on an (x, y) Grid}
\format{
A data frame illustrating the required raw data format for using
 a grid format to characterize a site layout. There are five
 columns: one giving turbine IDs (\code{turbine}), columns for the \code{x}
 and \code{y} coordinates on 1 m. grids that overlay the search plot, a
 column giving the carcass count in each grid cell, and a column giving the
 distance of each cell from the turbine. There is only one turbine, searched
 on road and pad out. Coordinates are in meters relative to the turbine at 
 (0, 0).
}
\usage{
layout_xy
}
\description{
Example Data for Site Layout on an (x, y) Grid
}
\keyword{datasets}
