% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lQint}
\alias{lQint}
\title{Integral in logQuadratic distribution}
\usage{
lQint(x, b1, b2)
}
\arguments{
\item{x}{non-negative numeric vector}

\item{a}{positive numeric vector}
}
\value{
scalar or vector of length = max(length(x), length(a)), with values
 of the shorter recycled to match the length of the longer a la pnorm etc.
}
\description{
Integral in logQuadratic distribution
}
\details{
The upper incomplete gamma function, following Wolfram Alpha, namely,
 incGam(x, a) = Gamma = integral(exp(-t) * t^(a - 1) dt from x to Inf),
 calculated using pgamma.
}
