% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mod_all}
\alias{mod_all}
\title{Names of All the Available Models}
\format{
An object of class \code{character} of length 17.
}
\usage{
mod_all
}
\description{
Names of All the Available Models
}
\keyword{datasets}
