% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mod_color}
\alias{mod_color}
\title{Vector of Colors Used in Graphs of Fitted Models}
\format{
An object of class \code{character} of length 17.
}
\usage{
mod_color
}
\description{
Vector of Colors Used in Graphs of Fitted Models
}
\keyword{datasets}
