% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mod_lty}
\alias{mod_lty}
\title{Vector of Line Types Used in Graphs of Fitted Models}
\format{
An object of class \code{numeric} of length 17.
}
\usage{
mod_lty
}
\description{
Vector of Line Types Used in Graphs of Fitted Models
}
\keyword{datasets}
