% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mod_offset}
\alias{mod_offset}
\title{Vector of GLM Offsets for Available Models}
\format{
An object of class \code{character} of length 17.
}
\usage{
mod_offset
}
\description{
Vector of GLM Offsets for Available Models
}
\keyword{datasets}
