% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mod_standard}
\alias{mod_standard}
\title{Vector of Names of Standard Models}
\format{
An object of class \code{character} of length 12.
}
\usage{
mod_standard
}
\description{
Vector of Names of Standard Models
}
\keyword{datasets}
