% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mod_xy}
\alias{mod_xy}
\title{Vector of Names of Models Available for Grid Layout}
\format{
An object of class \code{character} of length 8.
}
\usage{
mod_xy
}
\description{
Vector of Names of Models Available for Grid Layout
}
\keyword{datasets}
