% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_M.R
\name{MpriorOK}
\alias{MpriorOK}
\title{Check validity of format of custom prior for M}
\usage{
MpriorOK(prior)
}
\arguments{
\item{prior}{a custom prior for M must be a matrix with columns for M and
and associated probabalities P(M = m). The M column must begin at 0 and the
probabilities must sum to 1.}
}
\value{
boolean. Is the prior formatted properly?
}
\description{
Check validity of format of custom prior for M
}
