% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{natural}
\alias{natural}
\title{Vector of Names of Models with Natural Offset}
\format{
An object of class \code{logical} of length 17.
}
\usage{
natural
}
\description{
The natural offset is the area (m^2) searched at a given distance. Models
 that use the natural offset are referred to as "natural". Other models may
 use different offsets which alter the shape of the curve with distance in an
 a priori way that is unaffected by the data.
}
\keyword{datasets}
