% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{par_default}
\alias{par_default}
\title{Default Graphics Parameters}
\format{
An object of class \code{list} of length 65.
}
\usage{
par_default
}
\description{
Default Graphics Parameters
}
\keyword{datasets}
