% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{par_reset}
\alias{par_reset}
\title{reset par to default as in par_default for the \code{dwp} package}
\usage{
par_reset()
}
\description{
reset par to default as in par_default for the \code{dwp} package
}
