% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parm_name}
\alias{parm_name}
\title{List of Names of the Distribution Parameters Associated with Respective Models}
\format{
An object of class \code{list} of length 17.
}
\usage{
parm_name
}
\description{
List of Names of the Distribution Parameters Associated with Respective Models
}
\keyword{datasets}
