% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parOK}
\alias{parOK}
\title{Check Parameter Value Validity the Distribution}
\usage{
parOK(parms, distr)
}
\arguments{
\item{parms}{vector or matrix of named glm parameters (with \code{"r"} as the
distance variable)}

\item{distr}{name of the distribution}
}
\value{
vector (or scalar) of 0s, 1s, and 2s to indicate whether the parameters
 are non-extensible (i.e., flat-out bogus), valid for the distribution, or
 valid for the distribution and give finite point densities.
}
\description{
Performs a ouick check on whether the parameters given in \code{parms} are
 valid for the \code{distr}.
}
