% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{plot}
\alias{plot}
\alias{plot.ddArray}
\alias{plot.dd}
\title{plot dd and ddArray objects}
\usage{
\method{plot}{ddArray}(x, type = "CDF", extent = "full",
  distr = "all", xmax = NULL, resolution = 250,
  mod_highlight = NULL, par_reset = TRUE, ...)

\method{plot}{dd}(x, type = "CDF", extent = "full", xmax = NULL,
  resolution = 250, nsim = 1000, CL = 0.9, ...)
}
\arguments{
\item{x}{model(s) to plot}

\item{type}{Type or representation of carcass dispersion to plot:
\code{"CDF"}, \code{"PDF"}, or \code{"rcd"}. The \code{"CDF"} gives the
fraction of carcasses falling within \code{r} meters from a turbine and
\code{"PDF"} is the associated probability density. The \code{"rcd"} gives the
relative carcass density at a point \code{r} meters from a turbine and is
PDF/(2 * \pi * r).}

\item{extent}{Plot dispersions as fraction of total carcasses (\code{"full"})
or as fraction of carcasses within the searched area (\code{"win"}).}

\item{xmax}{maximum distance to show in the graph; if \code{xmax = NULL}, the
maximum distance is taken as the max distance in the data set to which the
models were fit.}

\item{resolution}{The number of line segments to break the curves into when
plotting (i.e., \code{x = seq(0, xmax, length.out = resolution)}). Higher
resolutions give smoother-looking curves.}

\item{mod_highlight}{Character string giving the name of the model to
highlight by plotting it last and with \code{lwd = 2}. If \code{NULL}, the
curve associated with the lowest (best) AICc score is highlighted.}

\item{nsim}{Number of simulation reps to use for estimating confidence bounds
for \code{dd} plot (ignored for \code{ddArray} objects)}

\item{CL}{confidence level to show in a \code{dd} plot (ignored for
\code{ddArray} objects)}

\item{set}{To show lines for all fitted models, use \code{set = "all"}; or
\code{set} = vector of model names to show lines for.}
}
\description{
Plot CDF, PDF, or rcd (relative carcass density) for a single
 carcass dispersion glm model (\code{dd} object) or a list of models
 (\code{ddArray} object).
}
\details{
\code{ddArray} objects are plotted with lines in order of decreasing
 AICc, so that the "better" models are closer to the top and more prominent.
 The model with the lowest AICc ("best" model) is plotted last with a heavier
 line than the others.

 \code{dd} plots the curve for the MLE of the parameters, along with a
 100\code{CL}\% confidence bounds determined for \code{nsim} simulation reps
}
