% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{prepmod}
\alias{prepmod}
\title{Internal Utility Function to Parse and Format Model for Calculating Psihat}
\usage{
prepmod(model, nsim)
}
\arguments{
\item{model}{\code{dd} or \code{ddSim} object}

\item{nsim}{number of simulation reps. If \code{nsim = 0}, return dd2ddSim}
}
\description{
Internal Utility Function to Parse and Format Model for Calculating Psihat
}
