% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{prepRing.simpleLayout}
\alias{prepRing.simpleLayout}
\title{Prepare Ring Data for Data Frames of Simple Turbine Layouts}
\usage{
\method{prepRing}{simpleLayout}(x)
}
\arguments{
\item{x}{a data frame with simple characterizations of turbines}

\item{cod}{carcass observation distances. Data frame with columns for
distances that carcasses were found at and which turbines. If \code{cod = NULL},
an empty template is created for carcass with descriptions of the turbines
(i.e., \code{r}, \code{exposure}, \code{ncarc}, \code{pinc}), and (optional)
\code{scVar}. A separate function can then be used to add carcasses distances later. This setup allows for easier
simulation of carcasses counts using the predefined template for turbines.
A "total" is added for the site...sums the carcasses and exposures}

\item{(optional)}{data frame or vector with carcass distances and turbine names}
}
\description{
Prepare Ring Data for Data Frames of Simple Turbine Layouts
}
