% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{prepRing_r}
\alias{prepRing_r}
\title{Tally carcasses by ring}
\usage{
prepRing_r(rdata, srad, rCol = "r")
}
\arguments{
\item{rdata}{data frame or array with a column giving the distances at which
carcasses were found. Alternatively, \code{rdata} may be a simple vector of
distances.}

\item{srad}{Search radius}

\item{rCol}{name of the column in \code{rdata} that includes the carcass
distances.}
}
\value{
Data frame with columns:
 \code{r} = the outer radius of 1m rings from 1 to \code{ceiling(srad)},
 \code{exposure} = area in ring = \code{(2 * pi * (r - 0.5))}
 \code{ncarc} = number of carcasses observed in each ring
}
\description{
Convert a vector of carcass distances into a data frame that
 gives a tally of the number of carcasses in 1m rings, appropriate for modeling
 carcass dispersion. Assumption is that all area within the search radius was
 searched and with uniform detection probability. Covariates associated with
 carcass counts are not allowed in this simplified data structure.
}
