% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{print.dd}
\alias{print.dd}
\title{Print dd and ddArray objects}
\usage{
\method{print}{dd}(x, ...)
}
\arguments{
\item{x}{a \code{dd} or \code{ddArray} object}

\item{...}{ignored}
}
\description{
Print dd and ddArray objects
}
