% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{print.dd}
\alias{print.dd}
\alias{print.ddArray}
\title{Print \code{dd} and \code{ddArray} Objects}
\usage{
\method{print}{dd}(x, ...)

\method{print}{ddArray}(x, ...)
}
\arguments{
\item{x}{a \code{\link[=ddFit]{ddArray}} or \code{\link[=ddFit]{ddArray}} object}

\item{...}{ignored}
}
\description{
\code{\link[=ddFit]{dd}} objects are lists consisting of all the 
 elements of \code{glm} objects and several additional elements related to 
 the distribution and its extensibility. Only a few of the elements are 
 printed automatically. Others elements of object \code{x} can be viewed and 
 extracted as with other lists in R, namely, by using the \code{x$element} or 
 \code{x[[element]]} operator, where \code{element} is the name of one of the 
 elements of \code{x} which can be viewed via \code{names(x)}.
}
