% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{readCarcass}
\alias{readCarcass}
\title{different types of data but all end up in the same place: list of arrays by
 (by turbine) with r, x, y, scVar}
\usage{
readCarcass(file_cod, unitCol = "turbine")
}
\arguments{
\item{rCol}{column for carcass distances from turbine}
}
\description{
different types of data but all end up in the same place: list of arrays by
 (by turbine) with r, x, y, scVar
}
