% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{readLayout}
\alias{readLayout}
\title{Create a shapeLayout}
\usage{
readLayout(file_layout, file_CO, file_turbine = NULL, unitCol = NULL)
}
\arguments{
\item{file_layout}{name of a shape file (polygons or multipolygons) that
delineates areas searched at each turbine. Search areas may be defined for
any number of search classes (associated with different detection
probabilities) and turbines. Search classes are listed the \code{scVar}
column and turbine IDs are specified in the (\code{unitCol}) column. If a
\code{unitCol} is provided (required if the site consists of more than one
turbine), each turbine should have a syntactically valid name. Search class
polygons at each turbine should be non-intersecting; otherwise, the function
exits with an error.}

\item{file_CO}{name of a shape file (points) for carcass observations
containing carcass discovery locations. If a \code{unitCol} name is provided,
turbine IDs must be provided for  carcass discoveries. Search classes are
optional.}

\item{file_turbine}{a shape file (points) giving the locations of the turbine
centers for turbines listed in the \code{unitCol} column in the \code{file_layout}}

\item{unitCol}{(optional) column name for turbine IDs. If a \code{unitCol} is
provided, a \code{file_turbine} with the turbine centers must also be
provided and the \code{unitCol} must be present in \code{file_turbine},
\code{file_layout}, and \code{file_CO}. Turbine IDs in the \code{unitCol}
must be syntactically valid R names (see Description above).}
}
\value{
A \code{shapeLayout} object, which is a list with spatial
 characteristics of the site, including
 \describe{
   \item{\code{$layout}}{turbine search area configurations (polygons and multipolygons)
     from \code{file_layout} shape file as an \code{sf} object.}
   \item{\code{$layoutAdj}}{polygons from \code{$layout} but all recentered at (0, 0)}
   \item{\code{$turbines}}{turbine centers (as \code{sf} object)}
   \item{\code{$carcasses}}{carcass discovery locations along with (optional)
     incidental information (turbine, visibility class, species,)}
   \item{\code{$unitCol, $ncarc, $tset, $tcenter}}{name of the column with turbine IDs
     (\code{$unitCol}, character string), number of carcasses found at each turbine
     (\code{$ncarc}, vector), turbine names (\code{$tset}, vector of character strings),
     locations of turbine centers (\code{$tcenter}, nturb by 2 array) UTMs of
     turbine centers, extracted and simplified from \code{$turbines}}
 }
}
\description{
Read shape files for search area polygons, turbine locations,
 and carcass discovery locations. If there is more than 1 turbine, there must
 be a turbine ID column (\code{unitCol}) in each of the files. The individual
 turbine ID's must be syntactically valid R names, i.e., contain combinations
 of letters, numbers, dot ( . ), and underscores ( _ ) only and not begin with
 a number or a dot followed by a number. Other characters are not allowed: no
 spaces, hyphens, parentheses, etc. All three shape files (search area
 polygons, turbine locations, carcass discovery locations) must have their
 three standard component files (.shp, .shx, .dbf) stored in the same folder.
}
