% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rmat}
\alias{rmat}
\title{simple utility function used in optimizing the truncated glm}
\usage{
rmat(r, distr)
}
\arguments{
\item{r}{vector of distances (>=0)}

\item{distr}{name of the distribution}
}
\value{
array with \code{length(r)} rows and p columns, where p is the number 
 of parameters in the glm (including the intercept). The first column is all 
 1s, and the remaining columns are functions of r, specifically, log(r), r, 
 r^2, r^3, or 1/r, depending on what the distribution requires.
}
\description{
simple utility function used in optimizing the truncated glm
}
