% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{stats}
\alias{stats}
\alias{stats.dd}
\title{Display a Table of Summary Statistics}
\usage{
stats(x, ...)

\method{stats}{dd}(x, extent = "full", zrad = 150, ...)
}
\arguments{
\item{x}{list of models (\code{ddArray}) or single model (\code{dd}) to calculate summary statistics for}

\item{...}{ignored}

\item{extent}{distributions within searched area (\code{"win"}) or extended beyond (\code{"full"})}

\item{zrad}{maximum distance that carcasses can lie
(only used when glm parameters not extensible to Inf)}
}
\value{
list (or list of lists if \code{x} is \code{ddArray}) with \code{$model}
 giving the model parameters and \code{$stats} giving the median, and 75th,
 90th, and 95th quantiles of carcass distances
}
\description{
Display a Table of Summary Statistics
}
