% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{[.ddArray}
\alias{[.ddArray}
\title{Subset set of fitted dispersion models (ddArray)}
\usage{
\method{[}{ddArray}(x, distr)
}
\arguments{
\item{x}{object to subset}

\item{distr}{vector of names of distributions to extract from fitted models}

\item{...}{ignored (but required format for CRAN)}
}
\value{
list of selected models with crucial statistics
}
\description{
Subset set of fitted dispersion models (ddArray)
}
\details{
Subset the ddArray object as if it were a simple vector
}
