% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{[.ddSim}
\alias{[.ddSim}
\title{Subset Simulated Dispersion Parameters while Preserving Attributes}
\usage{
\method{[}{ddSim}(x, i, j, ...)
}
\arguments{
\item{x}{object to subset}

\item{i, j}{row and column indices to subset}

\item{...}{ignored}
}
\value{
array with simulated beta parameters from the glm model, their
 conversion to distribution parameters. NOTE: subsetting to a column or a row
 returns a matrix rather than a vector. This simplifies the coding and makes
 it easier to maintain integrity of data structures, but behavior differs from
 what is done when subsetting standard R matrices and arrays to a single column
 or row. Also unlike with standard R arrays and matrices, the class structure
 and attributes are preserved upon subsetting.
}
\description{
Subset Simulated Dispersion Parameters while Preserving Attributes
}
\details{
Subset the ddSim object as if it were a simple matrix or array
}
