% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{tfit}
\alias{tfit}
\alias{tfit.dmod}
\title{fit a truncated distribution from a \code{dmod} object}
\usage{
tfit(x, ...)

\method{tfit}{dmod}(x, ...)
}
\arguments{
\item{x}{\code{dmod} object, which is a \code{glm} object from a Poisson
with the distance variable named "r", the search class named "svar", and
the name of the distribution added to the glm list objects as a character
string.}
}
\value{
a fitted distribution for carcass v distance
}
\description{
fit a truncated distribution from a \code{dmod} object
}
