% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xyr}
\alias{xyr}
\title{Locations of All Carcasses in Grid Data}
\format{
matrix with columns x, y, r for all 100 carcasses in the simulation to
 generate the carcass data for the xy grid that was searched on road and pad.
}
\usage{
xyr
}
\description{
Locations of All Carcasses in Grid Data
}
\keyword{datasets}
