% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{addCarcass}
\alias{addCarcass}
\alias{addCarcass.shapeCarcass}
\alias{addCarcass.data.frame}
\title{Add Carcasses to a Site Layout}
\usage{
addCarcass(x, ...)

\method{addCarcass}{shapeCarcass}(x, data_ring, plotLayout = NULL,
  ncarcReset = TRUE, ccCol = NULL, ...)

\method{addCarcass}{data.frame}(x, data_ring, ccCol = NULL,
  ncarcReset = TRUE, unitCol = "turbine", rCol = "r", ...)
}
\description{
Add Carcasses to a Site Layout
}
