% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{aic}
\alias{aic}
\alias{aic.ddArray}
\alias{aic.ddArraycc}
\alias{aic.dd}
\title{Calculate Akaike Information Criterion (AICc) for Distance Distributions}
\usage{
aic(x, ...)

\method{aic}{ddArray}(x, extent = "full", ...)

\method{aic}{ddArraycc}(x, extent = "full", ...)

\method{aic}{dd}(x, ...)
}
\arguments{
\item{x}{list of models (\code{\link[=ddFit]{ddArray}}) or single model 
(\code{\link[=ddFit]{dd}}) to calculate AICs for}

\item{...}{ignored}

\item{extent}{Include only the extensible models (\code{extent = "full"}) or
all models (\code{extent = "win"}), whether or not they can be extended 
beyond the search radius.}
}
\value{
Data frame with AICc and deltaAICc for all models in \code{x}
}
\description{
functions for calculating AICc for carcass dispersion
 models.
}
