% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{cof2parms}
\alias{cof2parms}
\alias{cof2parms.matrix}
\alias{cof2parms.numeric}
\alias{cof2parms.dmod}
\title{convert glm parameters into distr parameters}
\usage{
cof2parms(x, ...)

cof2parms(x, ...)

\method{cof2parms}{matrix}(x, distr, ...)

\method{cof2parms}{numeric}(x, distr, ...)

\method{cof2parms}{dmod}(x, ...)
}
\arguments{
\item{x}{structure containing the parameters values to be converted}

\item{...}{ignored [but required for CRAN]}

\item{distr}{name of the distribution}

\item{rvar}{name of the distance variable}

\item{x}{object (vector or matrix of parameters, dmod, or glm)
with named glm parameters ("r", "I(r^2)", "I(r^3)", "log(r)", or "I(1/r)")}

\item{distr}{name of the distribution}
}
\value{
array with columns for the glm parameters, the corresponding
 probability distribution parameters,

matrix of parameters
}
\description{
There is a natural connection between glm and exponential
 families of probability distributions. \code{cof2parms} converts the glm
 parameters into probability distribution parameters in cases where the
 glm converges to 0 as x approaches infininty and converges to x0 < Inf as
 x approaches 0.
}
