% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R, R/utilities.R
\name{cof2parms}
\alias{cof2parms}
\alias{cof2parms.matrix}
\alias{cof2parms.numeric}
\alias{cof2parms.dd}
\alias{parms2cof.matrix}
\title{Convert GLM Coefficients into Named Distribution Parameters}
\usage{
cof2parms(x, ...)

\method{cof2parms}{matrix}(x, distr, ...)

\method{cof2parms}{numeric}(x, distr, ...)

\method{cof2parms}{dd}(x, ...)

\method{parms2cof}{matrix}(x, distr, ...)
}
\arguments{
\item{x}{object (vector or matrix of parameters, dd, or glm)
with named glm parameters (\code{"r"}, \code{"I(r^2)"}, \code{"I(r^3)"}, 
\code{"log(r)"}, or \code{"I(1/r)"}). NOTE: This function has minimal 
error-checking.}

\item{...}{ignored}

\item{distr}{name of the distribution}
}
\value{
matrix of parameters
}
\description{
Convert GLM Coefficients into Named Distribution Parameters
}
