% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{ddCI}
\alias{ddCI}
\title{Calculate CI for CDF, PDF, or rcd for a \code{dd} model at distance \code{r}}
\usage{
ddCI(mod, r, type = "CDF", CL = 0.9, nsim = 1000, extent = "full",
  na.tol = 0.1)
}
\arguments{
\item{mod}{a \code{dd} object}

\item{r}{distance from turbine (scalar or vector)}

\item{type}{"CDF", "PDF", or "rcd"}

\item{CL}{confidence level for the confidence interval(s)}

\item{nsim}{number of simulation draws to base the estimate of CI on}

\item{na.tol}{maximum fraction of invalid parameter sets to discard when
constructing CIs; abort if \code{mean(mod[, "proper"]) > na.tol}}
}
\value{
array (\code{ddCI} class) with columns for distance and the CI bounds
}
\description{
Calculate CI for CDF, PDF, or rcd for a \code{dd} model at distance \code{r}
}
