% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{ddFit}
\alias{ddFit}
\alias{ddFit.data.frame}
\alias{ddFit.rings}
\alias{ddFit.list}
\alias{ddFit.xyLayout}
\title{Fit Distance Distribution Model(s)}
\usage{
ddFit(x, ...)

\method{ddFit}{data.frame}(x, distr = "standard", scVar = NULL,
  rCol = "r", expoCol = "exposure", ncarcCol = "ncarc",
  silent = FALSE, ...)

\method{ddFit}{rings}(x, distr = "standard", scVar = NULL,
  rCol = "r", expoCol = "exposure", ncarcCol = "ncarc",
  silent = FALSE, ...)

\method{ddFit}{list}(x, distr = "standard", scVar = NULL, rCol = "r",
  expoCol = "exposure", ncarcCol = "ncarc", silent = FALSE, ...)

\method{ddFit}{xyLayout}(x, distr = "standard", scVar = NULL,
  notSearched = NULL, rCol = "r", ncarcCol = "ncarc",
  unitCol = "turbine", silent = FALSE, ...)
}
\arguments{
\item{...}{ignored}

\item{scVar}{name of column in \code{rdat} with (optional) subdivisions
of rings into search classes (character strings)}

\item{rCol}{name of column in \code{rdat} with outer radius for 1m rings}

\item{expoCol}{name of "exposure" column of total area in each ring belonging
to each search class.}

\item{rdat}{Data frame with a tally of carcasses within 1 m rings, along
with outer radii of the rings, (optional) search class column, and the amount
of area searched in each search class and ring.}

\item{model}{names (vector of character strings) of glm distribution templates
to fit. Default is \code{"all"}, which fits \code{xep01}, \code{lognormal},
\code{xep1}, \code{xep12}, \code{xep123}, \code{xepi0},
\code{xep0123}, \code{xep012}, \code{xep2}, \code{MaxwellBoltzmann}, \code{constant},
\code{tnormal}, \code{exponential}, \code{xep0}, \code{chisq}, and
\code{inverse_gaussian}. Any subset of these may also be fit with a single
call to \code{ddFit}.}
}
\value{
A list of fitted glm models as \code{dd} objects in a \code{ddArray}
 object if a vector of distributions is fit or a single \code{dd} object if a
 single model is fit. The \code{dd} objects are lists with all the standard
 \code{\link[stats]{glm}} components, in addition to the following elements:
 \describe{
   \item{\code{$distr}}{name of the distribution ("xep01", etc.)}
   \item{\code{$parms}}{vector of distribution parameter estimates (or \code{NA}
     if the model based on the MLE is not extensible)}
   \item{\code{$varbeta}}{the variance-covariance matrix of the glm parameter
     estimates. NOTE: This is identical to the covariance matrix from the glm,
     which can be extracted via \code{summary(x)$cov.unscaled}}
   \item{\code{$scVar}}{name of the (optional) search class variable (or \code{NULL})}
   \item{\code{$ncarc}}{number of carcasses}
   \item{\code{$n}}{number of rings}
   \item{\code{$k}}{number of parameters}
   \item{\code{$srad}}{search radius}
 }
 When a \code{dd} object is printed, only a small subset of the elements are
 shown. To see a full list of the objects, use \code{names(x)}. The elements
 can be extracted in the usual R way via \code{$} or \code{[[x]]}.
}
\description{
Fit glm's for distance distribution models corresponding to
 xep01 (gamma), xep012, lognormal, xep1, xep12, xep123, truncated normal,
 exponential, xep2 (Rayleigh), xep0 (Pareto), xepi0 (inverse gamma), xep0123
 (normal gamma with x = tau), Maxwell Boltzmann, chi-squared, and/or inverse
 Gaussian distributions. By default, \code{ddFit} fits all the aforementioned
 models, but any subset of the models may be fit as well by using, for example,
 \code{model = exclude(c("xep123", "constant"))} to fit all except the
 \code{"xep123"} and \code{"constant"} models or \code{model = c("xep01", "lognormal")}
 to fit just the \code{"xep01"} and \code{"lognormal"} models.

 The glm is converted to a probability distribution by dividing by a
 normalizing constant, namely the integral of the glm evaluated from 0 to
 infinity. In some cases (most notably when the leading coefficient of the
 glm is positive so the fitted curve does not converge to zero as x increases),
 converted to a probability distribution. In these cases, the distribution
 parameters are given as \code{NA}.
}
