% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{ddFit}
\alias{ddFit}
\title{Fit Distance Distribution Model(s)}
\usage{
ddFit(ringData, model = "all", scVar = NULL, rCol = "r",
  expoCol = "exposure", ncarcCol = "ncarc")
}
\arguments{
\item{ringData}{Data frame with a tally of carcasses within 1 m rings, along
with outer radii of the rings, (optional) search class column, and the amount
of area searched in each search class and ring.}

\item{model}{names (vector of character strings) of glm distribution templates
to fit. Default is \code{"all"}, which fits \code{gamma}, \code{lognormal},
\code{logLinear}, \code{logQuadratic}, \code{logCubic}, \code{inverse_gamma},
\code{paranormal_gamma}, \code{Rayleigh}, \code{MaxwellBoltzmann}, \code{constant},
\code{tnormal}, \code{exponential}, \code{Pareto}, \code{chisq}, and
\code{inverse_gaussian}. Any subset of these may also be fit with a single
call to \code{ddFit}.}

\item{scVar}{name of column in \code{ringData} with (optional) subdivisions
of rings into search classes (character strings)}

\item{rCol}{name of column in \code{ringData} with outer radius for 1m rings}

\item{expoCol}{name of "exposure" column of total area in each ring belonging
to each search class.}
}
\description{
Fit glm's for distance distribution models corresponding to
 gamma, lognormal, log-linear, log-quadratic, log-cubic, truncated normal,
 exponential, Rayleigh, Pareto, inverse gamma, paranormal gamma, Maxwell
 Boltzmann, chi-squared, and/or inverse Gaussian distributions. By default,
 \code{ddFit} fits all the aforementioned models, but any subset of the models
 may be fit as well.

 The glm is converted to a probability distribution by dividing by a
 normalizing constant, namely the integral of the glm evaluated from 0 to
 infinity. In some cases (most notably when the leading coefficient of the
 glm is positive so the fitted curve does not converge to zero as x increases),
 converted to a probability distribution. In these cases, the distribution
 parameters are given as \code{NA}.
}
