% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{ddSim}
\alias{ddSim}
\alias{ddSim.dd}
\title{Simulation of Dispersion Parameters}
\usage{
ddSim(x, ...)

\method{ddSim}{dd}(x, nsim = 1000, extent = "full", ...)
}
\arguments{
\item{x}{object to simulate from}

\item{...}{ignored}

\item{nsim}{number of simulation draws}

\item{extent}{simulate according to full distribution, including extrapolation
beyond the search radius (\code{extent = "full"}); or restrict the
distribution to the area within the search radius (\code{extent = "win"}).}
}
\value{
array with simulated beta parameters from the glm model, and their
 conversion to distribution parameters
}
\description{
Simulation of Dispersion Parameters
}
