% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{estDWP}
\alias{estDWP}
\title{estDWP...obsolete?}
\usage{
estDWP(data, model, nboot = 1000)
}
\arguments{
\item{data}{data}

\item{model}{model}

\item{nboot}{nboot}
}
\value{
answer
}
\description{
estDWP...obsolete?
}
