% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{estpsi}
\alias{estpsi}
\alias{estpsi.data.frame}
\alias{estpsi.rpA}
\alias{estpsi.xyLayout}
\alias{estdwp.psiHat}
\title{Estimate Probability Carcass lands in Searched Area}
\usage{
estpsi(x, ...)

\method{estpsi}{data.frame}(x, model, extent = "full", nsim = 1000,
  zrad = 150, ...)

\method{estpsi}{rpA}(x, model, extent = "full", nsim = 1000,
  zrad = 150, ...)

\method{estpsi}{xyLayout}(x, model, extent = "full", nsim = 1000,
  zrad = 150, ...)

\method{estdwp}{psiHat}(x, ncarc, nboot = NULL, forGenEst = FALSE,
  silent = TRUE, ...)
}
\arguments{
\item{x}{data. Either \code{rdat} data frame for a single turbine, site,
or other entity, or a list of such entities.}

\item{...}{ignored}

\item{model}{A fitted \code{dd} model}

\item{extent}{calculate dwp within searched area only (\code{"win"}) or
for fulss complement of carcasses (\code{"full"}).}

\item{nsim}{number of parametric bootstrap iterations for estimating CIs}
}
\value{
\code{psiHat} object = list with
}
\description{
Estimated probability that carcass lands in searched area. This
 is an intermediate step in estimating dwp but is also interesting in its own
 right.
}
