% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{exclude}
\alias{exclude}
\title{remove particular distribution names from a longer list (for subsetting \code{ddArray})}
\usage{
exclude(what, from = mod_name)
}
\arguments{
\item{what}{vector of distribution names to exclude}

\item{from}{vector of distribution names to be excluded from}
}
\description{
remove particular distribution names from a longer list (for subsetting \code{ddArray})
}
