% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{off}
\alias{off}
\title{Utility Function for Constructing Offsets for GLMs}
\usage{
off(r, distr)
}
\arguments{
\item{r}{vector of distances}

\item{distr}{name of the distribution to calculate the offset for}
}
\value{
A vector of offset values to use with distances \code{r} when fitting
 the \code{distr} model.
}
\description{
This is a simple utility function for calculating offsets when exposure
 is assumed to be 100% at a given distance \code{r}. This is useful for
 calculating fitted distributions (\code{PDF} and \code{CDF}) but cannot be
 used in the fitting of the distributions themselves because it does not
 account for incomplete search coverages at given distances.
}
