% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R, R/fun0.R
\name{prepmod}
\alias{prepmod}
\title{Internal Utility Function to Parse and Format Model for Calculating psihat}
\usage{
prepmod(model, nsim)

prepmod(model, nsim)
}
\arguments{
\item{model}{\code{dd} or \code{ddSim} object}

\item{nsim}{number of simulation reps. If \code{nsim = 0}, return dd2ddSim}
}
\value{
\code{\link{ddSim}} object of simulated distribution model parameters

\code{\link{ddSim}} object of simulated distribution model parameters
}
\description{
Internal Utility Function to Parse and Format Model for Calculating psihat

Internal Utility Function to Parse and Format Model for Calculating Psihat
}
