% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rmat}
\alias{rmat}
\title{Simple Utility Function Used in Optimizing the GLM}
\usage{
rmat(r, distr)
}
\arguments{
\item{r}{vector of distances (>=0)}

\item{distr}{name of the distribution}
}
\value{
array with \code{length(r)} rows and p columns, where p is the number 
 of parameters in the glm (including the intercept). The first column is all 
 1s, and the remaining columns are functions of r, specifically, log(r), r, 
 r^2, r^3, or 1/r, depending on what the distribution requires.
}
\description{
A simple utility function that is used in fitting a GLM, creating a matrix
 of "x" values for use in the polynomial part of a xep-type model.
}
