% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{[.ddArray}
\alias{[.ddArray}
\title{Subset a Set of Fitted Dispersion Models (\code{ddArray})}
\usage{
\method{[}{ddArray}(x, distr)
}
\arguments{
\item{x}{object to subset}

\item{distr}{vector of names or integer indices of distributions to extract 
from fitted models}
}
\value{
list of selected models with crucial statistics
}
\description{
Subset a Set of Fitted Dispersion Models (\code{ddArray})
}
\details{
Subset the ddArray object as if it were a simple vector
}
